+++ USB_DEVICE/App/usbd_cdc_if.c
@@ -23,6 +23,7 @@
 
 /* USER CODE BEGIN INCLUDE */
 #include <string.h>
+#include "mcuforge_demo.h"
 
 /* USER CODE END INCLUDE */
 
@@ -265,7 +266,11 @@ static int8_t CDC_Control_HS(uint8_t cmd, uint8_t* pbuf, uint16_t length)
 static int8_t CDC_Receive_HS(uint8_t* Buf, uint32_t *Len)
 {
   /* USER CODE BEGIN 11 */
+#if MCUFORGE_DEMO_MODE
+  MCUForge_Demo_ReceiveBytes(Buf, *Len);
+#else
   (void)Len;
+#endif
   USBD_CDC_SetRxBuffer(&hUsbDeviceHS, &Buf[0]);
   USBD_CDC_ReceivePacket(&hUsbDeviceHS);
   return (USBD_OK);